services/backend_api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
httpx==0.28.1
google-cloud-bigquery==3.25.0
